+++ examples/client.js
@@ -3,7 +3,7 @@ var otText = require('ot-text');
 var CodeMirror = require('codemirror');
 var ShareDBCodeMirror = require('..');
 
-sharedb.types.register(otText.type);
+sharedb.types.map['json0'].registerSubtype(otText.type);
 
 var editorElem = document.getElementById('editor');
 var editor = CodeMirror.fromTextArea(editorElem, {
@@ -16,6 +16,7 @@ var shareConnection = new sharedb.Connection(socket);
 var doc = shareConnection.get('users', 'jane');
 
 ShareDBCodeMirror.attachDocToCodeMirror(doc, editor, {
+  key: 'content',
   verbose: true
 });