+++ test/headless-tests.js
@@ -7,7 +7,7 @@ var MockDoc = require('./mocks/mock-doc');
 function createCodeMirror(shareDoc, callback) {
   document.body.innerHTML = '<textarea id="editor"></textarea>';
   var codeMirror = CodeMirror.fromTextArea(document.getElementById('editor'));
-  ShareDBCodeMirror.attachDocToCodeMirror(shareDoc, codeMirror, callback);
+  ShareDBCodeMirror.attachDocToCodeMirror(shareDoc, codeMirror, {key: 'content'}, callback);
   return codeMirror;
 }
 
@@ -15,7 +15,7 @@ describe('Creation', function () {
   var doc, codeMirror;
 
   beforeEach(function (done) {
-    doc = new MockDoc('hi');
+    doc = new MockDoc({content: 'hi'});
     codeMirror = createCodeMirror(doc, done);
   });
 
@@ -25,13 +25,13 @@ describe('Creation', function () {
 
   describe('fresh doc', function () {
     beforeEach(function (done) {
-      doc = new MockDoc('');
+      doc = new MockDoc({});
       doc.type = null;
       codeMirror = createCodeMirror(doc, done);
     });
 
-    it('creates the doc as text', function () {
-      assert.equal('text', doc.type && doc.type.name);
+    it('creates the doc', function () {
+      assert.equal('json0', doc.type && doc.type.name);
     });
   });
 });
@@ -40,33 +40,33 @@ describe('CodeMirror edits', function () {
   var doc, codeMirror;
 
   beforeEach(function (done) {
-    doc = new MockDoc('');
+    doc = new MockDoc({content: ''});
     codeMirror = createCodeMirror(doc, done);
   });
 
   it('adds text', function () {
     var text = "aaaa\nbbbb\ncccc\ndddd";
     codeMirror.setValue(text);
-    assert.equal(text, doc.data);
+    assert.equal(text, doc.data.content);
   });
 
   it('adds empty text', function () {
     codeMirror.setValue('');
-    assert.equal('', doc.data);
+    assert.equal('', doc.data.content);
 
     codeMirror.setValue('a');
-    assert.equal('a', doc.data);
+    assert.equal('a', doc.data.content);
   });
 
   describe('with text in doc', function () {
     beforeEach(function (done) {
-      doc = new MockDoc('three\nblind\nmice\nsee\nhow\nthey\nrun\n');
+      doc = new MockDoc({content: 'three\nblind\nmice\nsee\nhow\nthey\nrun\n'});
       codeMirror = createCodeMirror(doc, done);
     });
 
     it('replaces a couple of lines', function () {
       codeMirror.replaceRange('evil\nrats\n', {line: 1, ch: 0}, {line: 3, ch: 0});
-      assert.equal('three\nevil\nrats\nsee\nhow\nthey\nrun\n', doc.data);
+      assert.equal('three\nevil\nrats\nsee\nhow\nthey\nrun\n', doc.data.content);
     });
   });
 });
@@ -75,44 +75,51 @@ describe('ShareJS changes', function () {
   var doc, codeMirror;
 
   beforeEach(function(done) {
-    doc = new MockDoc('');
+    doc = new MockDoc({content: ''});
     codeMirror = createCodeMirror(doc, done);
   });
 
   it('adds text', function () {
     var text = "aaaa\nbbbb\ncccc\ndddd";
-    doc.submitOp([text], false);
+    doc.submitOp([{p: ['content'], t: 'text', o: [text]}], false);
     assert.equal(text, codeMirror.getValue());
   });
 
   it('can edit a doc that has been empty', function () {
-    doc.submitOp([''], false);
+    doc.submitOp([{p: ['content'], t: 'text', o: ['']}], false);
     assert.equal('', codeMirror.getValue());
 
-    doc.submitOp(['a'], false);
+    doc.submitOp([{p: ['content'], t: 'text', o: ['a']}], false);
     assert.equal('a', codeMirror.getValue());
+  });
+
+  it('resets when the CodeMirror value and editor value diverge', function () {
+    doc.data = {content: 'foo'};
+    doc.submitOp([{p: ['content'], t: 'text', o: ['bar']}], false);
+
+    assert.equal('barfoo', codeMirror.getValue());
   });
 
   describe('with one line in the doc', function () {
     beforeEach(function (done) {
-      doc = new MockDoc('hi');
+      doc = new MockDoc({content: 'hi'});
       codeMirror = createCodeMirror(doc, done);
     });
 
     it('replaces a line', function () {
-      doc.submitOp([{d: 2}, 'hello'], false);
+      doc.submitOp([{p: ['content'], t: 'text', o: [{d: 2}, 'hello']}], false);
       assert.equal('hello', codeMirror.getValue());
     });
   });
 
   describe('with multiple lines in the doc', function() {
     beforeEach(function (done) {
-      doc = new MockDoc('three\nblind\nmice\nsee\nhow\nthey\nrun\n');
+      doc = new MockDoc({content: 'three\nblind\nmice\nsee\nhow\nthey\nrun\n'});
       codeMirror = createCodeMirror(doc, done);
     });
 
     it('replaces a couple of lines', function () {
-      doc.submitOp([6, {d: 11}, 'evil\nrats\n'], false);
+      doc.submitOp([{p: ['content'], t: 'text', o: [6, {d: 11}, 'evil\nrats\n']}], false);
       assert.equal('three\nevil\nrats\nsee\nhow\nthey\nrun\n', codeMirror.getValue());
     });
   });
@@ -122,10 +129,10 @@ describe('ShareDBCodeMirror', function() {
   var doc, codeMirror, shareDBCodeMirror;
 
   beforeEach(function (done) {
-    doc = new MockDoc('stuff');
+    doc = new MockDoc({content: 'stuff'});
     document.body.innerHTML = '<textarea id="editor"></textarea>';
     codeMirror = CodeMirror.fromTextArea(document.getElementById('editor'));
-    shareDBCodeMirror = ShareDBCodeMirror.attachDocToCodeMirror(doc, codeMirror, done);
+    shareDBCodeMirror = ShareDBCodeMirror.attachDocToCodeMirror(doc, codeMirror, {key: 'content'}, done);
   });
 
   describe('getValue', function () {
@@ -141,11 +148,11 @@ describe('ShareDBCodeMirror', function() {
 
     it('stops listening to CodeMirror changes', function () {
       codeMirror.setValue('blah');
-      assert.equal('stuff', doc.data);
+      assert.equal('stuff', doc.data.content);
     });
 
     it('ignores incoming ops', function () {
-      doc.submitOp([doc.data.length, ' more stuff'], false);
+      doc.submitOp([{p: ['content'], t: 'text', o: [doc.data.length, ' more stuff']}], false);
       assert.equal('stuff', codeMirror.getValue());
     });
   });