+++ History.md
@@ -1,3 +1,9 @@
+## 0.2.0 (2016-04-18)
+
+Switched to the json0 type, instead of text, for creating and manipulating docs.
+Ported over checks from share-codemirror that ensure the value in CodeMirror and
+the ShareDB doc don't diverge.
+
 ## 0.1.0 (2016-04-17)
 
 First release. Based on the ShareJS 0.7 CodeMirror bindings by Aslak Hellesøy.